# Don't render `styled-components` displayName in production