--- a/packages/app/babel.config.js
+++ b/packages/app/babel.config.js
@@ -1,0 +1,17 @@
+module.exports = function (api) {
+  return {
+    presets: ['next/babel'],
+    plugins: [
+      'lodash',
+      'date-fns',
+      [
+        'babel-plugin-styled-components',
+        {
+          ssr: true,
+          pure: true,
+          displayName: api.env('development'),
+        },
+      ],
+    ],
+  };
+};