## Summary

This saves between 2-20kB per HTML file and around ~2kB of JS per page. A small optimization per page but a bit larger for the whole export.

## Test results

### Before

**Bundle size**

```sh
Page                                                                        Size     First Load JS
┌ ● /                                                                       8.17 kB         597 kB
├   /_app                                                                   0 B             238 kB
├ ○ /404                                                                    818 B           273 kB
├ ● /500                                                                    811 B           273 kB
├ ● /actueel/gemeente/[code] (84227 ms)                                     4.39 kB         589 kB
├   ├ /nl/actueel/gemeente/GM0502 (432 ms)
├   ├ /nl/actueel/gemeente/GM0361 (431 ms)
├   ├ /nl/actueel/gemeente/GM0317 (427 ms)
├   ├ /nl/actueel/gemeente/GM1680 (419 ms)
├   ├ /nl/actueel/gemeente/GM0034 (408 ms)
├   ├ /nl/actueel/gemeente/GM0484 (390 ms)
├   ├ /nl/actueel/gemeente/GM0388 (388 ms)
├   └ [+345 more paths]
├ ● /actueel/veiligheidsregio/[code] (6401 ms)                              4.19 kB         589 kB
├   ├ /nl/actueel/veiligheidsregio/VR16 (375 ms)
├   ├ /nl/actueel/veiligheidsregio/VR18 (308 ms)
├   ├ /nl/actueel/veiligheidsregio/VR17 (301 ms)
├   ├ /nl/actueel/veiligheidsregio/VR01
├   ├ /nl/actueel/veiligheidsregio/VR02
├   ├ /nl/actueel/veiligheidsregio/VR03
├   ├ /nl/actueel/veiligheidsregio/VR04
├   └ [+18 more paths]
├ λ /api/choropleth/[...param]                                              0 B             238 kB
├ λ /api/choropleth/konva-node.d                                            0 B             238 kB
├ λ /api/choropleth/topojson-client.d                                       0 B             238 kB
├ λ /api/choropleth/topology                                                0 B             238 kB
├ λ /api/topo-json/[type]                                                   0 B             238 kB
├ ● /artikelen                                                              4.99 kB         277 kB
├ ● /artikelen/[slug] (3523 ms)                                             1.79 kB         573 kB
├   ├ /en/artikelen/houden-we-ons-aan-de-avondklok
├   ├ /nl/artikelen/houden-we-ons-aan-de-avondklok
├   ├ /en/artikelen/waarom-zijn-er-vaccins-op-voorraad
├   └ [+25 more paths]
├ ● /contact                                                                2.63 kB         571 kB
├ ● /gemeente                                                               2.16 kB         596 kB
├ ● /gemeente/[code] (58907 ms)                                             1.67 kB         595 kB
├   ├ /nl/gemeente/GM0197 (2196 ms)
├   ├ /nl/gemeente/GM1723 (2161 ms)
├   ├ /nl/gemeente/GM0482 (2105 ms)
├   ├ /nl/gemeente/GM0484 (2101 ms)
├   ├ /nl/gemeente/GM0141 (2099 ms)
├   ├ /nl/gemeente/GM0361 (2069 ms)
├   ├ /nl/gemeente/GM1680 (2048 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/positief-geteste-mensen (101113 ms)                    4.18 kB         613 kB
├   ├ /nl/gemeente/GM1680/positief-geteste-mensen (531 ms)
├   ├ /nl/gemeente/GM0308/positief-geteste-mensen (496 ms)
├   ├ /nl/gemeente/GM0203/positief-geteste-mensen (482 ms)
├   ├ /nl/gemeente/GM0482/positief-geteste-mensen (471 ms)
├   ├ /nl/gemeente/GM0200/positief-geteste-mensen (469 ms)
├   ├ /nl/gemeente/GM0202/positief-geteste-mensen (455 ms)
├   ├ /nl/gemeente/GM1901/positief-geteste-mensen (451 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/rioolwater (53645 ms)                                  1.67 kB         616 kB
├   ├ /nl/gemeente/GM0299/rioolwater (451 ms)
├   ├ /nl/gemeente/GM0867/rioolwater (437 ms)
├   ├ /nl/gemeente/GM0351/rioolwater (437 ms)
├   ├ /nl/gemeente/GM0627/rioolwater (409 ms)
├   ├ /nl/gemeente/GM1949/rioolwater (390 ms)
├   ├ /nl/gemeente/GM0626/rioolwater (388 ms)
├   ├ /nl/gemeente/GM0293/rioolwater (380 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/sterfte (98533 ms)                                     2.15 kB         611 kB
├   ├ /nl/gemeente/GM0059/sterfte (834 ms)
├   ├ /nl/gemeente/GM1680/sterfte (821 ms)
├   ├ /nl/gemeente/GM0482/sterfte (821 ms)
├   ├ /nl/gemeente/GM0358/sterfte (819 ms)
├   ├ /nl/gemeente/GM0034/sterfte (813 ms)
├   ├ /nl/gemeente/GM0613/sterfte (790 ms)
├   ├ /nl/gemeente/GM0197/sterfte (789 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/vaccinaties (739 ms)                                   3.04 kB         618 kB
├   ├ /nl/gemeente/GM1680/vaccinaties
├   ├ /nl/gemeente/GM0358/vaccinaties
├   ├ /nl/gemeente/GM0197/vaccinaties
├   └ [+349 more paths]
├ ● /gemeente/[code]/ziekenhuis-opnames (92650 ms)                          3.91 kB         612 kB
├   ├ /nl/gemeente/GM1954/ziekenhuis-opnames (492 ms)
├   ├ /nl/gemeente/GM0756/ziekenhuis-opnames (489 ms)
├   ├ /nl/gemeente/GM0197/ziekenhuis-opnames (477 ms)
├   ├ /nl/gemeente/GM1721/ziekenhuis-opnames (412 ms)
├   ├ /nl/gemeente/GM0034/ziekenhuis-opnames (391 ms)
├   ├ /nl/gemeente/GM0889/ziekenhuis-opnames (390 ms)
├   ├ /nl/gemeente/GM0664/ziekenhuis-opnames (383 ms)
├   └ [+345 more paths]
├ ● /internationaal                                                         2.94 kB         572 kB
├ ● /internationaal/positief-geteste-mensen                                 18.4 kB         603 kB
├ ● /internationaal/varianten                                               8.25 kB         601 kB
├ ● /landelijk                                                              690 B           592 kB
├ ● /landelijk/besmettelijke-mensen                                         5.49 kB         608 kB
├ ● /landelijk/brononderzoek                                                6.79 kB         617 kB
├ ● /landelijk/coronamelder                                                 2.61 kB         609 kB
├ ● /landelijk/gedrag                                                       4.65 kB         626 kB
├ ● /landelijk/gehandicaptenzorg                                            4.07 kB         610 kB
├ ● /landelijk/intensive-care-opnames                                       1.95 kB         612 kB
├ ● /landelijk/maatregelen                                                  941 B           605 kB
├ ● /landelijk/positief-geteste-mensen                                      4.61 kB         615 kB
├ ● /landelijk/reproductiegetal                                             2.54 kB         609 kB
├ ● /landelijk/rioolwater                                                   3.08 kB         615 kB
├ ● /landelijk/sterfte                                                      6.9 kB          613 kB
├ ● /landelijk/thuiswonende-ouderen                                         3.88 kB         610 kB
├ ● /landelijk/vaccinaties                                                  17.5 kB         630 kB
├ ● /landelijk/varianten                                                    1.15 kB         612 kB
├ ● /landelijk/verdenkingen-huisartsen                                      3.39 kB         610 kB
├ ● /landelijk/verpleeghuiszorg                                             4.07 kB         610 kB
├ ● /landelijk/ziekenhuis-opnames                                           3.51 kB         614 kB
├ ● /over                                                                   2.64 kB         571 kB
├ ● /over-risiconiveaus                                                     11.1 kB         580 kB
├ λ /sitemap.xml                                                            306 B           239 kB
├ ● /toegankelijkheid                                                       2.63 kB         571 kB
├ ● /veelgestelde-vragen                                                    1.22 kB         572 kB
├ ● /veiligheidsregio                                                       5.22 kB         609 kB
├ ● /veiligheidsregio/[code] (4981 ms)                                      1.63 kB         601 kB
├   ├ /nl/veiligheidsregio/VR01
├   ├ /nl/veiligheidsregio/VR02
├   ├ /nl/veiligheidsregio/VR03
├   └ [+22 more paths]
├ ● /veiligheidsregio/[code]/brononderzoek (11292 ms)                       5.5 kB          628 kB
├   ├ /nl/veiligheidsregio/VR23/brononderzoek (588 ms)
├   ├ /nl/veiligheidsregio/VR15/brononderzoek (564 ms)
├   ├ /nl/veiligheidsregio/VR04/brononderzoek (562 ms)
├   ├ /nl/veiligheidsregio/VR21/brononderzoek (492 ms)
├   ├ /nl/veiligheidsregio/VR25/brononderzoek (487 ms)
├   ├ /nl/veiligheidsregio/VR13/brononderzoek (486 ms)
├   ├ /nl/veiligheidsregio/VR20/brononderzoek (486 ms)
├   └ [+18 more paths] (avg 424 ms)
├ ● /veiligheidsregio/[code]/gedrag (8969 ms)                               1.71 kB         631 kB
├   ├ /nl/veiligheidsregio/VR11/gedrag (434 ms)
├   ├ /nl/veiligheidsregio/VR12/gedrag (424 ms)
├   ├ /nl/veiligheidsregio/VR14/gedrag (406 ms)
├   ├ /nl/veiligheidsregio/VR01/gedrag (405 ms)
├   ├ /nl/veiligheidsregio/VR15/gedrag (393 ms)
├   ├ /nl/veiligheidsregio/VR17/gedrag (392 ms)
├   ├ /nl/veiligheidsregio/VR16/gedrag (381 ms)
├   └ [+18 more paths] (avg 341 ms)
├ ● /veiligheidsregio/[code]/gehandicaptenzorg (13199 ms)                   2.7 kB          617 kB
├   ├ /nl/veiligheidsregio/VR03/gehandicaptenzorg (626 ms)
├   ├ /nl/veiligheidsregio/VR18/gehandicaptenzorg (619 ms)
├   ├ /nl/veiligheidsregio/VR17/gehandicaptenzorg (616 ms)
├   ├ /nl/veiligheidsregio/VR04/gehandicaptenzorg (613 ms)
├   ├ /nl/veiligheidsregio/VR05/gehandicaptenzorg (612 ms)
├   ├ /nl/veiligheidsregio/VR06/gehandicaptenzorg (602 ms)
├   ├ /nl/veiligheidsregio/VR07/gehandicaptenzorg (574 ms)
├   └ [+18 more paths] (avg 497 ms)
├ ● /veiligheidsregio/[code]/maatregelen (8048 ms)                          1.84 kB         613 kB
├   ├ /nl/veiligheidsregio/VR06/maatregelen (404 ms)
├   ├ /nl/veiligheidsregio/VR08/maatregelen (397 ms)
├   ├ /nl/veiligheidsregio/VR07/maatregelen (375 ms)
├   ├ /nl/veiligheidsregio/VR09/maatregelen (375 ms)
├   ├ /nl/veiligheidsregio/VR12/maatregelen (375 ms)
├   ├ /nl/veiligheidsregio/VR17/maatregelen (366 ms)
├   ├ /nl/veiligheidsregio/VR18/maatregelen (356 ms)
├   └ [+18 more paths]
├ ● /veiligheidsregio/[code]/positief-geteste-mensen (19622 ms)             2.45 kB         621 kB
├   ├ /nl/veiligheidsregio/VR06/positief-geteste-mensen (902 ms)
├   ├ /nl/veiligheidsregio/VR07/positief-geteste-mensen (901 ms)
├   ├ /nl/veiligheidsregio/VR03/positief-geteste-mensen (891 ms)
├   ├ /nl/veiligheidsregio/VR04/positief-geteste-mensen (880 ms)
├   ├ /nl/veiligheidsregio/VR08/positief-geteste-mensen (871 ms)
├   ├ /nl/veiligheidsregio/VR09/positief-geteste-mensen (866 ms)
├   ├ /nl/veiligheidsregio/VR05/positief-geteste-mensen (865 ms)
├   └ [+18 more paths] (avg 747 ms)
├ ● /veiligheidsregio/[code]/rioolwater (7765 ms)                           1.54 kB         621 kB
├   ├ /nl/veiligheidsregio/VR20/rioolwater (414 ms)
├   ├ /nl/veiligheidsregio/VR06/rioolwater (382 ms)
├   ├ /nl/veiligheidsregio/VR08/rioolwater (379 ms)
├   ├ /nl/veiligheidsregio/VR03/rioolwater (365 ms)
├   ├ /nl/veiligheidsregio/VR12/rioolwater (361 ms)
├   ├ /nl/veiligheidsregio/VR17/rioolwater (357 ms)
├   ├ /nl/veiligheidsregio/VR18/rioolwater (354 ms)
├   └ [+18 more paths]
├ ● /veiligheidsregio/[code]/risiconiveau (8342 ms)                         5.53 kB         620 kB
├   ├ /nl/veiligheidsregio/VR02/risiconiveau (463 ms)
├   ├ /nl/veiligheidsregio/VR04/risiconiveau (430 ms)
├   ├ /nl/veiligheidsregio/VR11/risiconiveau (384 ms)
├   ├ /nl/veiligheidsregio/VR08/risiconiveau (381 ms)
├   ├ /nl/veiligheidsregio/VR10/risiconiveau (376 ms)
├   ├ /nl/veiligheidsregio/VR09/risiconiveau (372 ms)
├   ├ /nl/veiligheidsregio/VR14/risiconiveau (370 ms)
├   └ [+18 more paths] (avg 309 ms)
├ ● /veiligheidsregio/[code]/sterfte (10302 ms)                             3.14 kB         617 kB
├   ├ /nl/veiligheidsregio/VR22/sterfte (502 ms)
├   ├ /nl/veiligheidsregio/VR10/sterfte (495 ms)
├   ├ /nl/veiligheidsregio/VR11/sterfte (493 ms)
├   ├ /nl/veiligheidsregio/VR20/sterfte (493 ms)
├   ├ /nl/veiligheidsregio/VR05/sterfte (458 ms)
├   ├ /nl/veiligheidsregio/VR07/sterfte (436 ms)
├   ├ /nl/veiligheidsregio/VR01/sterfte (432 ms)
├   └ [+18 more paths] (avg 389 ms)
├ ● /veiligheidsregio/[code]/thuiswonende-ouderen (10960 ms)                2.52 kB         617 kB
├   ├ /nl/veiligheidsregio/VR17/thuiswonende-ouderen (556 ms)
├   ├ /nl/veiligheidsregio/VR11/thuiswonende-ouderen (493 ms)
├   ├ /nl/veiligheidsregio/VR23/thuiswonende-ouderen (489 ms)
├   ├ /nl/veiligheidsregio/VR25/thuiswonende-ouderen (489 ms)
├   ├ /nl/veiligheidsregio/VR24/thuiswonende-ouderen (488 ms)
├   ├ /nl/veiligheidsregio/VR03/thuiswonende-ouderen (477 ms)
├   ├ /nl/veiligheidsregio/VR18/thuiswonende-ouderen (467 ms)
├   └ [+18 more paths] (avg 417 ms)
├ ● /veiligheidsregio/[code]/vaccinaties                                    3.04 kB         623 kB
├   ├ /nl/veiligheidsregio/VR01/vaccinaties
├   ├ /nl/veiligheidsregio/VR02/vaccinaties
├   ├ /nl/veiligheidsregio/VR03/vaccinaties
├   └ [+22 more paths]
├ ● /veiligheidsregio/[code]/verpleeghuiszorg (12301 ms)                    2.71 kB         617 kB
├   ├ /nl/veiligheidsregio/VR16/verpleeghuiszorg (588 ms)
├   ├ /nl/veiligheidsregio/VR22/verpleeghuiszorg (566 ms)
├   ├ /nl/veiligheidsregio/VR05/verpleeghuiszorg (565 ms)
├   ├ /nl/veiligheidsregio/VR01/verpleeghuiszorg (557 ms)
├   ├ /nl/veiligheidsregio/VR17/verpleeghuiszorg (554 ms)
├   ├ /nl/veiligheidsregio/VR07/verpleeghuiszorg (553 ms)
├   ├ /nl/veiligheidsregio/VR02/verpleeghuiszorg (550 ms)
├   └ [+18 more paths] (avg 465 ms)
├ ● /veiligheidsregio/[code]/ziekenhuis-opnames (8984 ms)                   3.93 kB         618 kB
├   ├ /nl/veiligheidsregio/VR01/ziekenhuis-opnames (542 ms)
├   ├ /nl/veiligheidsregio/VR06/ziekenhuis-opnames (530 ms)
├   ├ /nl/veiligheidsregio/VR07/ziekenhuis-opnames (525 ms)
├   ├ /nl/veiligheidsregio/VR16/ziekenhuis-opnames (461 ms)
├   ├ /nl/veiligheidsregio/VR05/ziekenhuis-opnames (416 ms)
├   ├ /nl/veiligheidsregio/VR08/ziekenhuis-opnames (381 ms)
├   ├ /nl/veiligheidsregio/VR15/ziekenhuis-opnames (374 ms)
├   └ [+18 more paths] (avg 320 ms)
├ ● /verantwoording                                                         1.22 kB         572 kB
└ ● /weekberichten/[slug] (2119 ms)                                         1.27 kB         573 kB
    ├ /nl/weekberichten/nieuwe-varianten-gooien-roet-in-het-eten (435 ms)
    ├ /en/weekberichten/nieuwe-varianten-gooien-roet-in-het-eten (357 ms)
    ├ /nl/weekberichten/minder-testen-door-winterweer (313 ms)
    ├ /en/weekberichten/minder-testen-door-winterweer
    ├ /en/weekberichten/kleine-landelijke-daling-maar-niet-in-alle-regio-s
    ├ /nl/weekberichten/kleine-landelijke-daling-maar-niet-in-alle-regio-s
    ├ /en/weekberichten/britse-variant-wint-terrein-in-nederland
    └ [+3 more paths]
+ First Load JS shared by all                                               238 kB
  ├ chunks/framework.56dc96.js                                              42.7 kB
  ├ chunks/main.fc0520.js                                                   24.3 kB
  ├ chunks/pages/_app.796170.js                                             169 kB
  ├ chunks/webpack.f9da33.js                                                2.07 kB
  └ css/fc7993b8fa7f45665623.css                                            329 B

λ  (Server)  server-side renders at runtime (uses getInitialProps or getServerSideProps)
○  (Static)  automatically rendered as static HTML (uses no initial props)
●  (SSG)     automatically generated as static HTML + JSON (uses getStaticProps)
   (ISR)     incremental static regeneration (uses revalidate in getStaticProps)

✨  Done in 233.56s.
✨  Done in 233.94s.
```

**HTML size**

| Page | HTML size  |
|---------------------------------------|------------|
| `/`                                   | `415.71KB` |
| `/landelijk/vaccinaties`              | `1.64MB`   |
| `/veiligheidsregio`                   | `60.86KB`  |
| `/veiligheidsregio/VR10/risiconiveau` | `893.33KB` |
| `/gemeente/GM0448/ziekenhuis-opnames` | `249.25KB` |
| `/over`                               | `41.51KB`  |


### After

**Bundle size**

```sh
Page                                                                        Size     First Load JS
┌ ● /                                                                       8.06 kB         596 kB
├   /_app                                                                   0 B             238 kB
├ ○ /404                                                                    818 B           272 kB
├ ● /500                                                                    811 B           272 kB
├ ● /actueel/gemeente/[code] (148052 ms)                                    4.38 kB         588 kB
├   ├ /nl/actueel/gemeente/GM0473 (1062 ms)
├   ├ /nl/actueel/gemeente/GM1896 (952 ms)
├   ├ /nl/actueel/gemeente/GM0355 (943 ms)
├   ├ /nl/actueel/gemeente/GM1701 (795 ms)
├   ├ /nl/actueel/gemeente/GM0880 (745 ms)
├   ├ /nl/actueel/gemeente/GM1969 (742 ms)
├   ├ /nl/actueel/gemeente/GM0642 (741 ms)
├   └ [+345 more paths] (avg 412 ms)
├ ● /actueel/veiligheidsregio/[code] (20531 ms)                             4.17 kB         588 kB
├   ├ /nl/actueel/veiligheidsregio/VR18 (1093 ms)
├   ├ /nl/actueel/veiligheidsregio/VR01 (950 ms)
├   ├ /nl/actueel/veiligheidsregio/VR07 (915 ms)
├   ├ /nl/actueel/veiligheidsregio/VR05 (911 ms)
├   ├ /nl/actueel/veiligheidsregio/VR11 (911 ms)
├   ├ /nl/actueel/veiligheidsregio/VR20 (893 ms)
├   ├ /nl/actueel/veiligheidsregio/VR13 (874 ms)
├   └ [+18 more paths] (avg 777 ms)
├ λ /api/choropleth/[...param]                                              0 B             238 kB
├ λ /api/choropleth/konva-node.d                                            0 B             238 kB
├ λ /api/choropleth/topojson-client.d                                       0 B             238 kB
├ λ /api/choropleth/topology                                                0 B             238 kB
├ λ /api/topo-json/[type]                                                   0 B             238 kB
├ ● /artikelen                                                              4.83 kB         276 kB
├ ● /artikelen/[slug] (5104 ms)                                             1.76 kB         572 kB
├   ├ /en/artikelen/houden-we-ons-aan-de-avondklok (323 ms)
├   ├ /nl/artikelen/houden-we-ons-aan-de-avondklok
├   ├ /en/artikelen/waarom-zijn-er-vaccins-op-voorraad
├   ├ /nl/artikelen/waarom-zijn-er-vaccins-op-voorraad
├   ├ /en/artikelen/test-artikel-plaatsen-afbeeldingen
├   ├ /nl/artikelen/test-artikel-plaatsen-afbeeldingen
├   ├ /en/artikelen/wat-zeggen-de-uitslagen-van-de-ggd-teststraten
├   └ [+21 more paths]
├ ● /contact                                                                2.63 kB         570 kB
├ ● /gemeente                                                               2.16 kB         594 kB
├ ● /gemeente/[code] (67134 ms)                                             1.67 kB         594 kB
├   ├ /nl/gemeente/GM0197 (2529 ms)
├   ├ /nl/gemeente/GM0141 (2517 ms)
├   ├ /nl/gemeente/GM0034 (2502 ms)
├   ├ /nl/gemeente/GM0059 (2465 ms)
├   ├ /nl/gemeente/GM0358 (2430 ms)
├   ├ /nl/gemeente/GM0484 (2398 ms)
├   ├ /nl/gemeente/GM0613 (2391 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/positief-geteste-mensen (117728 ms)                    4.12 kB         611 kB
├   ├ /nl/gemeente/GM0184/positief-geteste-mensen (799 ms)
├   ├ /nl/gemeente/GM0642/positief-geteste-mensen (659 ms)
├   ├ /nl/gemeente/GM0861/positief-geteste-mensen (618 ms)
├   ├ /nl/gemeente/GM1883/positief-geteste-mensen (607 ms)
├   ├ /nl/gemeente/GM0047/positief-geteste-mensen (598 ms)
├   ├ /nl/gemeente/GM0096/positief-geteste-mensen (576 ms)
…